requirements.txt
# Optional packages which may be used with REST framework

python_version>="3.5"
django
djangorestframework
